improvements to the bundle system
