+++ Packages/Basis Framework/Bundle Management/BasisLoadBundle.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 7ce824d514cb7654b95d7e6c2a502396
