+++ Packages/basisdk/Scripts/Editor/AssetBundleBuilder/BasisAssetBundlePipeline.cs
@@ -69,7 +69,7 @@ public static class BasisAssetBundlePipeline
             string prefabPath = TemporaryStorageHandler.SavePrefabToTemporaryStorage(prefab, settings, ref wasModified, out string uniqueID);
             string assetBundleName = AssetBundleBuilder.SetAssetBundleName(prefabPath, uniqueID, settings);
 
-           await AssetBundleBuilder.BuildAssetBundle(settings, assetBundleName, BasisBundleInformation, nameof(GameObject), Password);
+           await AssetBundleBuilder.BuildAssetBundle(settings, assetBundleName, BasisBundleInformation, "GameObject", Password);
             AssetBundleBuilder.ResetAssetBundleName(prefabPath);
             TemporaryStorageHandler.ClearTemporaryStorage(settings.TemporaryStorage);
             AssetDatabase.Refresh();
@@ -109,7 +109,7 @@ public static class BasisAssetBundlePipeline
             tempScenePath = TemporaryStorageHandler.SaveSceneToTemporaryStorage(scene, settings, out string uniqueID);
             string assetBundleName = AssetBundleBuilder.SetAssetBundleName(tempScenePath, uniqueID, settings);
 
-           await AssetBundleBuilder.BuildAssetBundle(settings, assetBundleName, BasisBundleInformation, nameof(Scene), Password);
+           await AssetBundleBuilder.BuildAssetBundle(settings, assetBundleName, BasisBundleInformation, "Scene", Password);
             TemporaryStorageHandler.ClearTemporaryStorage(tempScenePath);
 
             // Invoke the delegate after building the asset bundle
